BETA: All Fishes are now in XML instead of SQL '''(by nonom)'''.  * The new FishData class is using the Zoey's DocumentParser abstraction.  * New attributes, so now the fishes have the proper data.  * New fishes and categories that actually are missing.  * Renamed all the old attributes matching the retail names, like groups and grades.  * Added javadocs, cleaned and formatted.
'''NOTE: Require DP [DP8783]'''

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/model/StatsSet.java
@@ -393,7 +393,7 @@ public final class StatsSet
 	 * @param deflt : float designating the default value if value associated with the key is null
 	 * @return double : value associated to the key
 	 */
-	public double getDouble(String name, float deflt)
+	public double getDouble(String name, double deflt)
 	{
 		Object val = _set.get(name);
 		if (val == null)
